Removing zero on hand inventory records. closes #11

+++ src/main/java/guru/sfg/beer/inventory/service/services/AllocationServiceImpl.java
@@ -72,6 +72,9 @@ public class AllocationServiceImpl implements AllocationService {
                 beerOrderLine.setQuantityAllocated(allocatedQty + inventory);
                 beerInventory.setQuantityOnHand(0);
 
+                //beerInventoryRepository.delete(beerInventory);
+            }
+            if(beerInventory.getQuantityOnHand()==0){
                 beerInventoryRepository.delete(beerInventory);
             }
         });
